Just use char (basic_string<char>) and assume file encoding is UTF-8
In the future it would be best to add a translation layer between
u8strings and filesystem or Qt. On a file system any binary file name
is always correct (excepting the NUL char), so such layer is needed.
Qt in the user-interface assumes UTF-8, so there will be really no
translation at all.

But there's known bug in Qt that it doesn't handle correctly Linux file names
because Qt wrongly assumes they're Unicode (or encoded with something that's
specified in an enviroment variable of a particular process):
<https://bugreports.qt.io/browse/QTBUG-59402>

It's unlikely that Qt fixes it since now it's simply too late.
Since Qt won't be able to show all files, then perhaps showing
a warning to the user (when Xefis is first run) would be an appropriate
idea? Or maybe just ignore this problem? It's not essential after all…