Disconnect signal for updating items when deleting SimulatorWidget
When an object in the QStackedWidget gets deleted, it notifies parent
(QStackedWidget) about this, and QStackedWidget changes currently
displayed item to another. It emits currentItemChanged. In SimulatorWidget
that signal was connected to a function that did something with editors,
particularly with editor that was just deleted.

So when destroying SimulatorWidget, start with disconnecting signal
QStackedWidget::currentItemChanged.